Move Image and PixelFormat to net_traits.
