+++ components/compositing/compositor.rs
@@ -18,9 +18,10 @@ use gleam::gl::types::{GLint, GLsizei};
 use image::{DynamicImage, ImageFormat, RgbImage};
 use ipc_channel::ipc::{self, IpcSender, IpcSharedMemory};
 use ipc_channel::router::ROUTER;
-use msg::constellation_msg::{Image, PixelFormat, Key, KeyModifiers, KeyState};
+use msg::constellation_msg::{Key, KeyModifiers, KeyState};
 use msg::constellation_msg::{PipelineId, PipelineIndex, PipelineNamespaceId};
 use msg::constellation_msg::{TraversalDirection, WindowSizeType};
+use net_traits::image::base::{Image, PixelFormat};
 use profile_traits::mem::{self, Reporter, ReporterRequest};
 use profile_traits::time::{self, ProfilerCategory, profile};
 use script_traits::{AnimationState, AnimationTickType, ConstellationControlMsg};
