+++ components/compositing/compositor_thread.rs
@@ -9,7 +9,8 @@ use compositor::CompositingReason;
 use euclid::point::Point2D;
 use euclid::size::Size2D;
 use ipc_channel::ipc::IpcSender;
-use msg::constellation_msg::{Image, Key, KeyModifiers, KeyState, PipelineId};
+use msg::constellation_msg::{Key, KeyModifiers, KeyState, PipelineId};
+use net_traits::image::base::Image;
 use profile_traits::mem;
 use profile_traits::time;
 use script_traits::{AnimationState, ConstellationMsg, EventResult};
