+++ components/msg/Cargo.toml
@@ -14,7 +14,6 @@ bitflags = "0.7"
 cssparser = {version = "0.7", features = ["heap_size", "serde-serialization"]}
 heapsize = "0.3.0"
 heapsize_plugin = "0.1.2"
-ipc-channel = "0.5"
 plugins = {path = "../plugins"}
 serde = "0.8"
 serde_derive = "0.8"
