+++ components/net_traits/lib.rs
@@ -34,6 +34,7 @@ extern crate serde_derive;
 extern crate url;
 extern crate util;
 extern crate uuid;
+extern crate webrender_traits;
 extern crate websocket;
 
 use cookie_rs::Cookie;
